fstab: increase zRAM to 384 MiB
A bit more memory working space is nice on a memory constrained device,
as it reduce low memory killing and should improve stability. Thanks to
the quad core processor, the performance impact should be a bit less than
it is condor. This is a happy medium between Motorola's choice of 256 MiB
and Qualcomm's choice of 512 MiB for 1GB RAM devices.