otus: don't get stuck in charge only mode
Technically, the proper way to do this would be restructure the
charge only mode initialization to match 8226-common, so that the
proper USB VID/PID pair are set for charge only mode. This is a
minor edge case that I don't care about too much, and the 8226 way
is a bit overly complicated (requiring kernel additions etc).

Nobody complained about charge only mode the way it was before, so
let's just make it the way it used to be.

Change-Id: I558a604eced8c8d181bd015ffb083e7aaa4cf035